appSkeleton: add messages system

+++ RWParcerCore/Application/UseCases/UserService/FavoritesService/RemoveFromFavoritesUseCase.cs
@@ -11,9 +11,10 @@ namespace RWParcerCore.Application.UseCases.UserService.FavoritesService
 
         public async Task RemoveFromFavoritesAsync(string userId, TrainVO train)
         {
+            if (await _userRepository.IsUserBannedAsync(userId)) throw new UnauthorizedAccessException($"User {userId} is banned");
             if (!await _userRepository.IsUserRegistredAsync(userId)) throw new KeyNotFoundException($"User with ID {userId} not found");
 
-            var favoriteToRemove = (await _favoriteRepository.GetFavoritesAsync(userId)).FirstOrDefault(f => f.Train.Equals(train)) ?? throw new KeyNotFoundException("Favorite item not found");
+            var favoriteToRemove = (await _favoriteRepository.GetFavoritesAsync(userId)).FirstOrDefault(f => f.Train.Equals(train)) ?? throw new KeyNotFoundException($"Favorite item not found {userId}");
 
             await _favoriteRepository.RemoveFavoriteAsync(favoriteToRemove);
         }
